EXPA-06 UPD: popup menu refactored
